+++ packages/entrypoint/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sonata-api/entrypoint
 
+## 0.0.3
+
+### Patch Changes
+
+- Make router metadata available to the runtime
+- Updated dependencies
+  - @sonata-api/types@0.1.12
+
 ## 0.0.2
 
 ### Patch Changes